requirements.txt
absl-py==1.4.0
accelerate==0.18.0
asttokens==2.2.1
astunparse==1.6.3
av==10.0.0
backcall==0.2.0
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
docutils==0.19
executing==1.2.0
fast-ml==3.68
filelock==3.12.0
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
imageio==2.28.0
imageio-ffmpeg==0.4.8
ipykernel==6.22.0
ipython==8.13.1
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.12.0
Kivy==2.2.0.dev0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
kivy-deps.sdl2-dev==0.6.0
Kivy-examples==2.2.0.dev0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.1
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.5.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.15.1
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
torch==2.0.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
